+++ provider/mem/mem.go
@@ -29,7 +29,6 @@ import (
 	"go.opentelemetry.io/otel/trace"
 
 	"github.com/prometheus/alertmanager/eventrecorder"
-	"github.com/prometheus/alertmanager/eventrecorder/eventrecorderpb"
 	"github.com/prometheus/alertmanager/featurecontrol"
 	"github.com/prometheus/alertmanager/provider"
 	"github.com/prometheus/alertmanager/store"
@@ -348,7 +347,7 @@ func (a *Alerts) Put(ctx context.Context, alerts ...*types.Alert) error {
 		a.callback.PostStore(alert, existing)
 
 		if !existing {
-			a.recorder.RecordEvent(ctx, func() *eventrecorderpb.EventData {
+			a.recorder.RecordEvent(ctx, func() eventrecorder.EventData {
 				return eventrecorder.NewAlertCreatedEvent(alert)
 			})
 		}